# improve ngram performance with JIT