Improve ngram performance by 20%.

random_hv(10,10000)
99.1 µs ± 1.63 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
78 µs ± 1.05 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)