@mikeheddes 2022-05-17
Hi, thank you for the PR, good results. However, I don't think the library is at a stage to start micro-optimizing all the functions. If you have an improvement that is still easy to maintain then that would be much appreciated but this improvement takes away from maintainability and understandability which we prioritize higher than efficiency.